requirements.txt
flask
pygame
whisper
pydub
RPi.GPIO
numpy
